move csv_to_sys_param into SystemParameters class, including tests
