+++ geojson_modelica_translator/system_parameters/system_parameters.py
@@ -30,6 +30,9 @@ OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 import json
 import os
+from pathlib import Path
+from copy import deepcopy
+import pandas as pd
 
 from jsonpath_ng.ext import parse
 from jsonschema.validators import _LATEST_VERSION as LatestValidator
@@ -204,3 +207,86 @@ class SystemParameters(object):
             results.append(error.message)
 
         return results
+
+    def _parse_items(self, measure_folder: Path) -> None:
+        """
+        Go through each folder (OpenStudio measure) in a feature and pull out the mass-flow-rate data file & modelica loads file
+        """
+        if str(measure_folder).endswith('_export_time_series_modelica'):
+            self.measure_list.append(Path(measure_folder) / "building_loads.csv")
+        elif str(measure_folder).endswith('_export_modelica_loads'):
+            self.measure_list.append(Path(measure_folder) / "modelica.mos")
+
+    def _parse_feature(self, feature: Path) -> None:
+        """
+        Go through each feature directory in UO SDK output
+        """
+        for item in feature.iterdir():
+            if item.is_dir():
+                self._parse_items(item)
+
+    def csv_to_sys_param(self, scenario_dir, feature_file, sys_param_filename, overwrite=True):
+        if not Path(scenario_dir).exists():
+            raise Exception(f"Unable to find your scenario. The path you provided was: {scenario_dir}")
+
+        if not Path(feature_file).exists():
+            raise Exception(f"Unable to find your feature file. The path you provided was: {feature_file}")
+
+        if Path(sys_param_filename).exists() and not overwrite:
+            raise Exception(f"Output file already exists and overwrite is False: {sys_param_filename}")
+
+        sys_param_template = Path(__file__).parent / 'time_series_template.json'
+        # sys_param_template = self.data
+
+        self.measure_list = []
+
+        # Parse the sys_param template
+        with open(sys_param_template) as template_file:
+            param_template = json.load(template_file)
+
+        # Grab filepaths from sdk output
+        for thing in scenario_dir.iterdir():
+            if thing.is_dir():
+                self._parse_feature(thing)
+
+        # Parse the FeatureFile
+        building_ids = []
+        with open(feature_file) as json_file:
+            data = json.load(json_file)
+            for feature in data['features']:
+                if not feature['properties']['type'] == 'Site Origin':
+                    building_ids.append(feature['properties']['id'])
+
+        # Make sys_param template entries for each feature_id
+        building_list = []
+        for building in building_ids:
+            feature_info = deepcopy(param_template['buildings']['custom'][0])
+            feature_info['geojson_id'] = building
+            building_list.append(feature_info)
+
+        # Grab the modelica file for the each Feature, and add it to the appropriate building dict
+        district_nominal_mfrt = 0
+        for building in building_list:
+            building_nominal_mfrt = 0
+            for measure_file_path in self.measure_list:
+                feature_name, measure_load_name = str(measure_file_path).split('/')[-3:-1]
+                if feature_name != building['geojson_id']:
+                    continue
+                if (measure_file_path.suffix == '.mos'):
+                    building['load_model_parameters']['time_series']['filepath'] = str(measure_file_path)
+                if (measure_file_path.suffix == '.csv') and ('_export_time_series_modelica' in str(measure_load_name)):
+                    mfrt_df = pd.read_csv(measure_file_path)
+                    building_nominal_mfrt = mfrt_df['massFlowRateHeating'].max()
+                    building['load_model_parameters']['time_series']['nominal_flow_building'] = float(building_nominal_mfrt)
+                district_nominal_mfrt += building_nominal_mfrt
+
+        # Remove buildings that don't have successful simulations, with modelica outputs
+        building_list = [x for x in building_list if x['load_model_parameters']['time_series']['filepath'] is not None]
+        if len(building_list) == 0:
+            raise Exception("No Modelica files found. The UO SDK simulations may not have been successful")
+
+        param_template['buildings']['custom'] = building_list
+        param_template['buildings']['default']['ets_model_parameters']['indirect']['nominal_flow_district'] = district_nominal_mfrt
+
+        with open(sys_param_filename, 'w') as outfile:
+            json.dump(param_template, outfile, indent=2)
