+++ geojson_modelica_translator/system_parameters/time_series_template.json
