+++ tests/system_parameters/data/sdk_output_skeleton/example_project.json
